Added a dummy retirement trends by beneficiary table
