+++ carbonmark-data/components/charts/helpers/DataTable/configurations/KlimaRetirementsByBeneficiaryListConfiguration.tsx
@@ -2,12 +2,12 @@ import { RawRetirementsItem } from "lib/charts/types";
 import KlimaRetirementsListConfigurationBase from "./KlimaRetirementsListConfigurationBase";
 import { Columns } from "./types";
 
-export default class KlimaRetirementsByPoolListConfiguration extends KlimaRetirementsListConfigurationBase {
+export default class KlimaRetirementsByBeneficiaryListConfiguration extends KlimaRetirementsListConfigurationBase {
   getColumns(): Columns<RawRetirementsItem> {
     return {
       beneficiary: this.beneficiaryColumn,
       project_id: this.projectIdColumn,
-      token: this.tokenColumn,
+      token: this.protocolColumn,
       retirement_date: this.retirementDateColumn,
       quantity: this.quantityColumn,
       transaction_id: this.transactionIdColumn,
