+++ carbonmark-data/components/charts/helpers/DataTable/configurations/KlimaRetirementsByTokenListConfiguration.tsx
@@ -1,22 +1,16 @@
-import { t } from "@lingui/macro";
 import { RawRetirementsItem } from "lib/charts/types";
-import layout from "theme/layout.module.scss";
-import KlimaRetirementsByTokenListConfigurationBase from "./KlimaRetirementsByListConfigurationBase";
+import KlimaRetirementsListConfigurationBase from "./KlimaRetirementsListConfigurationBase";
 import { Columns } from "./types";
 
-export default class KlimaRetirementsByTokenListConfiguration extends KlimaRetirementsByTokenListConfigurationBase {
+export default class KlimaRetirementsByTokenListConfiguration extends KlimaRetirementsListConfigurationBase {
   getColumns(): Columns<RawRetirementsItem> {
-    const columns = super.getColumns();
-    columns["token"] = {
-      header: t`Token`,
-      cellStyle: layout.textCenter,
-      dataKey: "token",
-      formatter: (x: string | number) => {
-        if (x == "BCT" || x == "NCT") return "TCO2";
-        if (x == "UBO" || x == "NBO") return "C3T";
-        return "";
-      },
+    return {
+      beneficiary: this.beneficiaryColumn,
+      project_id: this.projectIdColumn,
+      token: this.protocolColumn,
+      retirement_date: this.retirementDateColumn,
+      quantity: this.quantityColumn,
+      transaction_id: this.transactionIdColumn,
     };
-    return columns;
   }
 }
